+++ app/Dtos/QuestionCategoryDto.php
@@ -7,22 +7,40 @@ class QuestionCategoryDto {
     /***
      * @var string $name
      */
-    protected $name;
+    protected string $name;
 
 
     /***
      * @var string $slug
      */
-    protected $slug;
+    protected string $slug;
 
 
+    /**
+     * @var string $description
+     */
+    protected string  $description;
 
-    public function __construct(string $name, string $slug)
+    /**
+     * @var string $image
+     */
+    protected string $image;
+
+
+    /**
+     * @param string $name
+     * @param string $description
+     * @param string $image
+     * @param string $slug
+     */
+    public function __construct(string $name, string $description, string $image, string $slug)
     {
         $this->name = $name;
+        $this->description = $description;
+        $this->image = $image;
         $this->slug = $slug;
-    }
 
+    }
 
 
     /***
@@ -40,6 +58,22 @@ class QuestionCategoryDto {
     public function getSlug() : string
     {
         return $this->slug;
+    }
+
+    /**
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return $this->description;
+    }
+
+    /**
+     * @return string
+     */
+    public function getImage(): string
+    {
+        return $this->image;
     }
 
 }