+++ app/Models/Answer.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Models\Traits\LikeTrait;
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -9,7 +10,7 @@ use Illuminate\Database\Eloquent\Relations\MorphTo;
 
 class Answer extends Model
 {
-    use CrudTrait;
+    use CrudTrait, LikeTrait;
 
     /*
     |--------------------------------------------------------------------------
@@ -31,6 +32,14 @@ class Answer extends Model
     |--------------------------------------------------------------------------
     */
 
+    public function formattedDate(){
+        return $this->created_at->format('d-m-Y H:m');
+    }
+
+    public function answeredByCurrentUser(User $user) : bool
+    {
+        return $this->user->id == $user->id;
+    }
     /*
     |--------------------------------------------------------------------------
     | RELATIONS