+++ app/Services/AnswerService.php
@@ -16,11 +16,10 @@ class AnswerService {
     {
         $answer = new Answer();
 
-        $answer->body = $request->getBody();
-        $answer->image = $request->getImage();
+        $answer->answer = $request->getBody();
         $answer->user_id = $request->getUserId();
-        $answer->question_id = $request->getQuestionId();
-        $answer->parent_id = $request->getParentId();
+        $answer->answerable_type = $request->getAnswerableType();
+        $answer->answerable_id = $request->getAnswerableId();
 
         if(!$answer->save())
         {