+++ database/migrations/2021_10_06_175333_create_question_category_table.php
@@ -17,6 +17,7 @@ class CreateQuestionCategoryTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug')->unique();
+            $table->text('description');
             $table->timestamps();
         });
     }