+++ resources/views/answerquestion/category_with_questions.blade.php
@@ -1,0 +1,9 @@
+@extends('layout.app')
+
+@section('content')
+<div clas="category">
+    <div class="max-with">
+        <x-questions-section :questions="$questions" :text="'Саволҳо аз ' . $category->name"></x-questions-section>
+    </div>
+</div>
+@endsection