+++ resources/views/answerquestion/show.blade.php
@@ -85,15 +85,15 @@
                                 @endif
                                     @if(Auth::check() && $question->askedByCurrentUser(Auth::user()))
                                         <div class="d-flex">
-                                            <form action="{{ route('questions.destroy', $question->id) }}" method="POST">
+                                            <form action="{{ route('answers.destroy', $question->id) }}" method="POST">
                                                 @csrf
                                                 @method("DELETE")
                                                 <button class="like-btn danger m-3"><i class="far fa-trash-alt"></i></button>
                                             </form>
-                                            <form action="{{ route('questions.edit', $question->id) }}" >
-                                                @csrf
-                                                <button class="like-btn m-3"><i class="far fa-edit"></i></button>
-                                            </form>
+{{--                                            <form action="{{ route('answers.edit', $question->id) }}" >--}}
+{{--                                                @csrf--}}
+{{--                                                <button class="like-btn m-3"><i class="far fa-edit"></i></button>--}}
+{{--                                            </form>--}}
                                         </div>
                                     @endif
                             </div>
@@ -118,75 +118,25 @@
                             @csrf
                             <div class="form-item">
                                 <label class="input-label" for="answer"><h3>Ҷавоби шумо:</h3></label>
-                                <textarea name="body" id="summernote" class="textarea" cols="30" placeholder="Ҷавобатонро инҷо нависед... " rows="7"></textarea>
+                                @error('answer')
+                                <p class="red">{{ $message }}</p>
+                                @enderror
+                                <textarea name="answer" id="summernote" class="textarea" cols="30" placeholder="Ҷавобатонро инҷо нависед... " rows="7"></textarea>
                             </div>
-                            <input type="hidden" name="question_id" value="{{ $question->id }}">
+
+                            <input type="hidden" name="answerable_id" value="{{ $question->id }}">
+                            <input type="hidden" name="answerable_type" value="question">
                             <button class="btn-b" id="create-btn" data-auth="login" type="submit">Ҷавоб додан</button>
                         </form>
                     </div>
+
+
+
+                    <x-answers-section :answers="$question->answers"></x-answers-section>
                 </div>
             </section>
 
 
-            {{-- answers --}}
-            <br>
-            <br>
-            <section class="max-width">
-                <div class="answers">
-                    <h2>Ҷавобҳо</h2>
-                    @forelse($question->answers as $answer)
-                        <div class="answer p-7">
-                            <div class="name-img d-flex align-center">
-                                <img src="{{ asset('avatars/' . $answer->user->avatar_path) }}"  width="50" alt="">
-                                <h4 class="ml-4"><a href="#">{{$answer->user->name}} {{$answer->user->lastname}}</a></h4>
-                            </div>
-                            <div class="p-5 content">
-                                <p>{!! $answer->body !!}</p>
-                            </div>
-                            <div class="icons">
-                                {{-- like section --}}
-                                @if(Auth::check())
-                                    @if($answer->likes->contains('user_id', Auth::user()->id))
-                                        <a class="like-btn green" href="/like/remove/answer/{{ $answer->id }}">
-                                            <i class="fas fa-thumbs-up"></i>
-                                            <span class="like-count">{{ $answer->likes->count() }}</span>
-                                        </a>
-                                    @else
-                                        <a class="like-btn" href="/like/answer/{{ $answer->id }}">
-                                            <i class="fas fa-thumbs-up"></i>
-                                            <span class="like-count">{{ $answer->likes->count() }}</span>
-                                        </a>
-                                    @endif
-
-                                    @if($answer->dislikes->contains('user_id', Auth::user()->id))
-                                        <a class="like-btn red" href="/like/remove/answer/{{ $question->id }}">
-                                            <i class="fas fa-thumbs-down"></i>
-                                            <span class="like-count">{{ $answer->dislikes->count() }}</span>
-                                        </a>
-                                    @else
-                                        <a class="like-btn" href="/dislike/answer/{{ $question->id }}">
-                                            <i class="fas fa-thumbs-down"></i>
-                                            <span class="like-count">{{ $answer->dislikes->count() }}</span>
-                                        </a>
-                                    @endif
-                                @else
-                                    <a class="like-btn not-allow">
-                                        <i class="fas fa-thumbs-up"></i>
-                                        <span class="like-count">{{ $answer->likes->count() }}</span>
-                                    </a>
-                                    <a class="like-btn not-allow">
-                                        <i class="fas fa-thumbs-down"></i>
-                                        <span class="like-count">{{ $answer->dislikes->count() }}</span>
-                                    </a>
-                                @endif
-                                <a class="answer-btn" href="{{ route('questions') }}/{{$question->id}}">Ҷавоб додан</a>
-                            </div>
-                        </div>
-                    @empty
-                        <h3>Аввалин шуда савол гузоред!</h3>
-                    @endforelse
-                </div>
-            </section>
 
 @endsection